+++ examples/JuMP Model by Hand - MLD-Block.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.13
+# v0.19.18
 
 using Markdown
 using InteractiveUtils